+++ ft_puthex.c
@@ -6,14 +6,14 @@
 /*   By: padan-pe <padan-pe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/16 18:37:43 by padan-pe          #+#    #+#             */
-/*   Updated: 2025/02/18 18:59:37 by padan-pe         ###   ########.fr       */
+/*   Updated: 2025/02/18 19:16:01 by padan-pe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "ft_printf.h"
 
-int	ft_puthex(unsigned long x, char *base)//%x%X
+int	ft_puthex(size_t x, char *base)//%x%X
 {
 	int		count;
